Build: add navbar
